Reorganized structure to fit server-client architecture
