--- config/__init__.py
@@ -1,3 +1,0 @@
-from .constants import DISPLAY, MAP, TILE, PLAYER, NETWORK
-
-__all__ = ['DISPLAY', 'MAP', 'TILE', 'PLAYER', 'NETWORK']
