--- core/__init__.py
@@ -1,4 +1,0 @@
-from .game import Game
-from .input_manager import InputManager
-
-__all__ = ['Game', 'InputManager']
