--- entities/__init__.py
@@ -1,4 +1,0 @@
-from .map import Map
-from .player import Player
-
-__all__ = ['Map', 'Player']
