--- graphics/__init__.py
@@ -1,3 +1,0 @@
-from .raycasting import RayCasting
-
-__all__ = ['RayCasting']
