--- main.py
@@ -1,10 +1,0 @@
-from core import Game
-
-
-def main():
-    game = Game()
-    game.run()
-
-
-if __name__ == "__main__":
-    main()
