--- network/__init__.py
@@ -1,3 +1,0 @@
-from .server import Server
-
-__all__ = ['Server']
