+++ server/network/__init__.py
@@ -1,0 +1,3 @@
+from .game_server import GameServer
+
+__all__ = ['GameServer']
