+++ shared/__init__.py
@@ -1,0 +1,3 @@
+from .constants import NETWORK
+
+__all__ = ['NETWORK']
