+++ shared/constants.py
@@ -1,0 +1,4 @@
+NETWORK = {
+    "SERVER_PORT": 5003,
+    "ALLOWED_CONNECTIONS": 2,
+}
