--- ui/__init__.py
@@ -1,3 +1,0 @@
-from .menu import Menu
-
-__all__ = ['Menu']
